+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * openapi: fix detection of default (undefined) schema implementation
+* server/openapi: fix analyzing of class-decorated handlers for OpenAPI generation
 
 # 1.7.2
 * server: HttpServer now accepts connections inside of workerPool, not single thread - for better scalability under very heavy load